+++ CHANGELOG.md
@@ -5,6 +5,13 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+## [0.8.0] - 2025-12-15
+- Add `Source` improvements:
+  - Add `Source` display page
+  - Add `Source` highlight support for Rust, C, and C++ using [arborium](arborium.bearcove.eu)(treesitter).
+- Send child process stderr to null
+- Resolve info addresses with another symbol address lookup in `Symbol` page.
+- Poll slower when not expecting gdb response
 
 ## [0.7.0] - 2025-11-14
 - Hexdump memory mapping selection using `H` [#192](https://github.com/wcampbell0x2a/heretek/pull/192)